+++ modular_citadel/code/modules/mob/living/silicon/robot/dogborg_equipment.dm
@@ -309,9 +309,7 @@ SLEEPER CODE IS IN game/objects/items/devices/dogborg_sleeper.dm !
 	var/leap_at
 	var/disabler
 	var/laser
-	var/sleeper_g
-	var/sleeper_r
-	var/sleeper_nv
+	var/sleeper_state = BORGBELLY_NONE
 
 #define MAX_K9_LEAP_DIST 4 //because something's definitely borked the pounce functioning from a distance.
